+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file is used to list changes made in each version of envbuilder.
 
+## 0.2.0:
+
+* Default to using `node.chef_environment` if `node['ENV']`` is not set
+  ([@prognostikos](https://github.com/prognostikos))
+* Add the ability to use an encrypted data bag
+  ([@prognostikos](https://github.com/prognostikos))
+* Add the ability to configure the permissions/mode of the generated file
+  ([@prognostikos](https://github.com/prognostikos))
+
 ## 0.1.0:
 
 * Initial release of envbuilder